various fixes for how group messages are displayed
