+++ src/groups.h
@@ -37,6 +37,13 @@ typedef struct groupchat {
 
 
 void group_init(GROUPCHAT *g, uint32_t group_number, _Bool av_group);
-void group_add_message(GROUPCHAT *g, int peer_id, const uint8_t *message, size_t length);
+
+void group_add_message(GROUPCHAT *g, int peer_id, const uint8_t *message, size_t length, uint8_t m_type);
+
+void group_peer_add(GROUPCHAT *g, uint32_t peer_id, _Bool our_peer_number);
+
+void group_peer_del(GROUPCHAT *g, uint32_t peer_id);
+
 void group_peer_name_change(GROUPCHAT *g, uint32_t peer_id, const uint8_t *name, size_t length);
+
 void group_free(GROUPCHAT *g);
